ajout de la fonctionnalité pour sauvegarder dans un fichier

+++ ProjetSGBD/src/modele/type/ByteSGBD.java
@@ -25,7 +25,7 @@ public class ByteSGBD extends TypeSGBD<Byte> {
 	}
 
 	@Override
-	public int taille() {
+	public long taille() {
 		return 1;
 	}
