+++ ProjetSGBD/src/modele/type/TypeSGBD.java
@@ -9,6 +9,6 @@ public abstract class TypeSGBD<T> {
 	public abstract void serialisation(DataOutputStream os, T data) throws IOException;
 	public abstract T deserialisation(DataInputStream is) throws IOException;
 	public abstract T parse(String data);
-	public abstract int taille();
+	public abstract long taille();
 	public String toString(T data) {return data.toString();}
 }
